requirements.txt
django
djangorestframework
markdown
django-filter
djangorestframework_simplejwt
psycopg2
Pillow
